+++ src/pages/EditorPage/EditorPage.tsx
