+++ src/pages/Layout/Layout.tsx
@@ -1,12 +1,12 @@
 import Cookies from 'js-cookie';
 import { useContext, useEffect } from 'react';
 import { Outlet } from 'react-router-dom';
-import { ApiClient } from '../api/api-client';
-import { UserProfileBody } from '../api/types';
-import { Footer } from '../components/Footer/Footer';
-import { Header } from '../components/Header/Header';
-import AuthContext from '../context/auth-context';
-import { CookieNames } from '../shared/constants';
+import { ApiClient } from '../../api/api-client';
+import { UserProfileBody } from '../../api/types';
+import { Footer } from '../../components/Footer/Footer';
+import { Header } from '../../components/Header/Header';
+import AuthContext from '../../context/auth-context';
+import { CookieNames } from '../../shared/constants';
 
 export function Layout() {
   const { isAuthenticated, setAuth } = useContext(AuthContext);
